put resources list functionality in one file, streamlined BLM
